Made a Pawn class, and Rules interface
The rules interface adds additional rules that need to be inherited by
all the other "rules" interfaces. The pawn class extends Piece and 
implements PawnRules and defines a new method called transform. This
method allows the pawn to transform into another piece once it reaches
the other side of the board.